Refactored how spellchecking is called
Also fixed style in several places.
