+++ packages/test-middlewares/src/middlewares.js
@@ -17,8 +17,7 @@
 module.exports = {
   in: [
     async (context, next, done) => {
-      const { user, brain, userMessage, config } = context; // eslint-disable-line no-unused-vars
-      const muted = await brain.userGet(user, '_isMuted');
+      const muted = await context.brain.userGet(context.user, '_isMuted');
       if (muted) {
         await done();
       } else {
@@ -28,8 +27,7 @@ module.exports = {
   ],
   out: [
     async (context, next) => {
-      const { user, brain, botMessages, config, userMessage } = context; // eslint-disable-line no-unused-vars
-      await brain.userSet(user, 'isOutMiddlewareWorking', true);
+      await context.brain.userSet(context.user, 'isOutMiddlewareWorking', true);
       await next();
     },
   ],
